+++ Modules/BandPassComputeComplex/src/BandPassComputeComplex.cpp
@@ -154,19 +154,19 @@ bool BandPassComputeComplex::run() {
 
 void BandPassComputeComplex::setParam(const std::string& paramName, const std::any& value) {
     if (paramName == "sample rate") {
-        m_sampleRate = static_cast<double>(std::any_cast<int32_t>(value));
+        m_sampleRate = static_cast<double>(std::any_cast<int64_t>(value));
         DEBUG << "BandPassComputeComplex sample rate set to: " << m_sampleRate << std::endl;
         return;
     }
 
     if (paramName == "filter order") {
-        m_filterOrder = std::any_cast<int32_t>(value);
+        m_filterOrder = std::any_cast<int64_t>(value);
         DEBUG << "BandPassComputeComplex filter order set to: " << m_filterOrder << std::endl;
         return;
     }
 
     if (paramName == "block size") {
-        (void)std::any_cast<int32_t>(value);
+        (void)std::any_cast<int64_t>(value);
         DEBUG << "BandPassComputeComplex::setParam: 'block size' is accepted for compatibility and ignored." << std::endl;
         return;
     }
